trip_be/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
freezegun==1.2.2
geographiclib==2.0
geopy==2.3.0
idna==3.4
iniconfig==2.0.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
